move invariant manager and run make fmt
