+++ service/worker/scanner/executions/invariants/invariantManager_test.go
@@ -20,7 +20,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-package shard
+package invariants
 
 import (
 	"testing"
